Refactor CourseStepFragment data loading to CoursesRepository
Moved the data loading logic for course steps from `CourseStepFragment` to `CoursesRepository`.
Introduced `CourseStepData` model to hold the step data.
Updated `CoursesRepositoryImpl` to implement the new method.
Refactored `CourseStepFragment` to use the repository method and removed local data classes.
