+++ app/src/main/java/org/ole/planet/myplanet/model/CourseStepData.kt
@@ -1,0 +1,8 @@
+package org.ole.planet.myplanet.model
+
+data class CourseStepData(
+    val step: RealmCourseStep,
+    val resources: List<RealmMyLibrary>,
+    val stepExams: List<RealmStepExam>,
+    val stepSurvey: List<RealmStepExam>
+)
