+++ app/src/main/java/org/ole/planet/myplanet/ui/courses/CourseStepFragment.kt
@@ -17,8 +17,8 @@ import org.ole.planet.myplanet.MainApplication
 import org.ole.planet.myplanet.R
 import org.ole.planet.myplanet.base.BaseContainerFragment
 import org.ole.planet.myplanet.databinding.FragmentCourseStepBinding
+import org.ole.planet.myplanet.model.CourseStepData
 import org.ole.planet.myplanet.model.RealmCourseStep
-import org.ole.planet.myplanet.model.RealmMyCourse.Companion.isMyCourse
 import org.ole.planet.myplanet.model.RealmMyLibrary
 import org.ole.planet.myplanet.model.RealmStepExam
 import org.ole.planet.myplanet.model.RealmUser
@@ -31,21 +31,6 @@ import org.ole.planet.myplanet.utils.CameraUtils.ImageCaptureCallback
 import org.ole.planet.myplanet.utils.CameraUtils.capturePhoto
 import org.ole.planet.myplanet.utils.MarkdownUtils.prependBaseUrlToImages
 import org.ole.planet.myplanet.utils.MarkdownUtils.setMarkdownText
-
-private data class CourseStepData(
-    val step: RealmCourseStep,
-    val resources: List<RealmMyLibrary>,
-    val stepExams: List<RealmStepExam>,
-    val stepSurvey: List<RealmStepExam>,
-    val userHasCourse: Boolean
-)
-
-private data class IntermediateStepData(
-    val step: RealmCourseStep,
-    val resources: List<RealmMyLibrary>,
-    val stepExams: List<RealmStepExam>,
-    val stepSurvey: List<RealmStepExam>
-)
 
 @AndroidEntryPoint
 class CourseStepFragment : BaseContainerFragment(), ImageCaptureCallback {
@@ -97,35 +82,7 @@ class CourseStepFragment : BaseContainerFragment(), ImageCaptureCallback {
     }
 
     private suspend fun loadStepData(): CourseStepData {
-        val intermediateData = databaseService.withRealmAsync { realm ->
-            val step = realm.where(RealmCourseStep::class.java)
-                .equalTo("id", stepId)
-                .findFirst()
-                ?.let { realm.copyFromRealm(it) }!!
-            val resources = realm.where(RealmMyLibrary::class.java)
-                .equalTo("stepId", stepId)
-                .findAll()
-                .let { realm.copyFromRealm(it) }
-            val stepExams = realm.where(RealmStepExam::class.java)
-                .equalTo("stepId", stepId)
-                .equalTo("type", "courses")
-                .findAll()
-                .let { realm.copyFromRealm(it) }
-            val stepSurvey = realm.where(RealmStepExam::class.java)
-                .equalTo("stepId", stepId)
-                .equalTo("type", "surveys")
-                .findAll()
-                .let { realm.copyFromRealm(it) }
-            IntermediateStepData(step, resources, stepExams, stepSurvey)
-        }
-        val userHasCourse = coursesRepository.isMyCourse(user?.id, intermediateData.step.courseId)
-        return CourseStepData(
-            step = intermediateData.step,
-            resources = intermediateData.resources,
-            stepExams = intermediateData.stepExams,
-            stepSurvey = intermediateData.stepSurvey,
-            userHasCourse = userHasCourse
-        )
+        return coursesRepository.getCourseStepData(stepId!!)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -137,6 +94,8 @@ class CourseStepFragment : BaseContainerFragment(), ImageCaptureCallback {
                 resources = data.resources
                 stepExams = data.stepExams
                 stepSurvey = data.stepSurvey
+
+                val userHasCourse = coursesRepository.isMyCourse(user?.id, step.courseId)
 
                 fragmentCourseStepBinding.btnResources.text =
                     getString(R.string.resources_size, resources.size)
@@ -155,7 +114,7 @@ class CourseStepFragment : BaseContainerFragment(), ImageCaptureCallback {
                 fragmentCourseStepBinding.description.movementMethod =
                     LinkMovementMethod.getInstance()
 
-                if (!data.userHasCourse) {
+                if (!userHasCourse) {
                     fragmentCourseStepBinding.btnTakeTest.visibility = View.GONE
                     fragmentCourseStepBinding.btnTakeSurvey.visibility = View.GONE
                 }
@@ -179,7 +138,7 @@ class CourseStepFragment : BaseContainerFragment(), ImageCaptureCallback {
                         textWithSpans.removeSpan(urlSpan)
                     }
                 }
-                if (isVisible && data.userHasCourse) {
+                if (isVisible && userHasCourse) {
                     launchSaveCourseProgress()
                 }
             }
